--- lib/device_type_enum.dart
@@ -1,3 +1,0 @@
-part of responsive_builder;
-
-enum DeviceType { Mobile, Tablet, Desktop, Watch }
