+++ lib/responsive_files/device_type_enum.dart
@@ -1,0 +1,2 @@
+part of '../responsive_builder.dart';
+enum DeviceType { Mobile, Tablet, Desktop, Watch }
