+++ pubspec.yaml
@@ -1,6 +1,6 @@
 name: responsive_builder
 description: A new Flutter package project.
-version: 0.0.1
+version: 0.1.0
 homepage:
 
 environment:
